+++ src/main/java/org/opensearch/flowframework/template/ProcessNode.java
@@ -17,33 +17,66 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
+/**
+ * Representation of a process node in a workflow graph.  Tracks predecessor nodes which must be completed before it can start execution.
+ */
 public class ProcessNode {
     private final String id;
-    private CompletableFuture<String> future;
+    private CompletableFuture<String> future = null;
 
     // will be populated during graph parsing
     private Set<ProcessNode> predecessors = Collections.emptySet();
 
+    /**
+     * Create this node with a unique id.
+     */
     ProcessNode(String id) {
         this.id = id;
     }
 
+    /**
+     * Returns the node's id.
+     * @return the node id.
+     */
     public String getId() {
         return id;
     }
 
+    /**
+     * Returns a {@link CompletableFuture} if this process is executing.
+     * Relies on the node having been sorted and executed in an order such that all predecessor nodes have begun execution first (and thus populated this value).
+     *
+     * @return A future indicating the processing state of this node.
+     * Returns {@code null} if it has not begun executing, should not happen if a workflow is sorted and executed topologically.
+     */
     public CompletableFuture<String> getFuture() {
         return future;
     }
 
+    /**
+     * Returns the predecessors of this node in the workflow.
+     * The predecessor's {@link #getFuture()} must complete before execution begins on this node.
+     *
+     * @return a set of predecessor nodes, if any.  At least one node in the graph must have no predecessors and serve as a start node.
+     */
     public Set<ProcessNode> getPredecessors() {
         return predecessors;
     }
 
-    public void setPredecessors(Set<ProcessNode> predecessors) {
+    /**
+     * Sets the predecessor node.  Called by {@link TemplateParser}.
+     *
+     * @param predecessors The predecessors of this node.
+     */
+    void setPredecessors(Set<ProcessNode> predecessors) {
         this.predecessors = Set.copyOf(predecessors);
     }
 
+    /**
+     * Execute this node in the sequence. Initializes the node's {@link CompletableFuture} and completes it when the process completes.
+     *
+     * @return this node's future. This is returned immediately, while process execution continues asynchronously.
+     */
     public CompletableFuture<String> execute() {
         this.future = new CompletableFuture<>();
         CompletableFuture.runAsync(() -> {
@@ -60,16 +93,17 @@ public class ProcessNode {
                 return;
             }
             System.out.println(">>> Starting " + this.id);
-            sleep(id.contains("ingest") ? 8000 : 4000);
+            // TODO: Here is where we would call out to workflow step API
+            workflowExecute(this.id);
             System.out.println("<<< Finished " + this.id);
             future.complete(this.id);
         });
         return this.future;
     }
 
-    private void sleep(long i) {
+    private void workflowExecute(String s) {
         try {
-            Thread.sleep(i);
+            Thread.sleep(s.contains("ingest") ? 8000 : 4000);
         } catch (InterruptedException e) {}
     }
