+++ src/main/java/org/opensearch/flowframework/template/ProcessSequenceEdge.java
@@ -10,19 +10,38 @@ package org.opensearch.flowframework.template;
 
 import java.util.Objects;
 
+/**
+ * Representation of an edge between process nodes in a workflow graph.
+ */
 public class ProcessSequenceEdge {
     private final String source;
     private final String destination;
 
+    /**
+     * Create this edge with the id's of the source and destination nodes.
+     *
+     * @param source The source node id.
+     * @param destination The destination node id.
+     */
     ProcessSequenceEdge(String source, String destination) {
         this.source = source;
         this.destination = destination;
     }
 
+    /**
+     * Gets the source node id.
+     *
+     * @return the source node id.
+     */
     public String getSource() {
         return source;
     }
 
+    /**
+     * Gets the destination node id.
+     *
+     * @return the destination node id.
+     */
     public String getDestination() {
         return destination;
     }
